Crate tynm

Returns type names with a specifiable number of module segments as a String.

// === std library === //
assert_eq!(
    std::any::type_name::<Option<String>>(),
    "core::option::Option<alloc::string::String>",
);

// === tynm === //
// Simple type name:
assert_eq!(tynm::type_name::<Option<String>>(), "Option<String>",);

// Type name with 1 module segment, starting from the most significant module.
assert_eq!(
    tynm::type_namem::<Option<String>>(1),
    "core::..::Option<alloc::..::String>",
);

// Type name with 1 module segment, starting from the least significant module.
assert_eq!(
    tynm::type_namen::<Option<String>>(1),
    "..::option::Option<..::string::String>",
);

// Type name with 1 module segment from both the most and least significant modules.
#[rustfmt::skip]
mod rust_out { pub mod two { pub mod three { pub struct Struct; } } }
assert_eq!(
    tynm::type_namemn::<rust_out::two::three::Struct>(1, 1),
    "rust_out::..::three::Struct",
);

Motivation

The std::any::type_name function stabilized in Rust 1.38 returns the fully qualified type name with all module segments. This can be difficult to read in error messages, especially for type-parameterized types.

Often, the simple type name is more readable, and enough to distinguish the type referenced in an error.

Structs

TypeNameDisplay
Helper struct for printing type names directly to format!.

Enums

TypeName
Organizes type name string into distinct parts.

Functions

type_name
Returns the simple type name.

type_namem
Returns the type name with at most m most significant module path segments.

type_namemn
Returns the type name with m most significant, and n least significant module path segments.

type_namen
Returns the type name with at most n least significant module path segments.