Function tynm::type_namem

pub fn type_namem<T>(m: usize) -> String

Returns the type name with at most m most significant module path segments.

Parameters

m: Number of most significant module path segments to include.

Type Parameters

T: Type whose simple type name should be returned.

Examples

assert_eq!(
    tynm::type_namem::<Option<String>>(1),
    "core::..::Option<alloc::..::String>",
);